requirements.txt
py-clob-client
websockets
python-dotenv
asyncpg
aiohttp
web3